Abstract database from bot core
The bot core was directly importing and depending on the Postgresdb
implementation. Created an abstract database class and split the
existing database class into a Postgres implementation and a test
dictionary implementation.

Additionaly, fixed a bug regarding the name not displaying on the
leaderboard correctly when a user didn't have a nickname.